packages: move ConfigReader and config types to separate config package
