+++ packages/config/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: [require.resolve('@backstage/cli/config/eslint')],
+  rules: {
+    'jest/expect-expect': 0,
+  },
+};
