+++ packages/config/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@backstage/config",
+  "description": "Config API used by Backstage core, backend, and CLI",
+  "version": "0.1.1-alpha.6",
+  "private": false,
+  "publishConfig": {
+    "access": "public"
+  },
+  "homepage": "https://backstage.io",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/spotify/backstage",
+    "directory": "packages/config"
+  },
+  "keywords": [
+    "backstage"
+  ],
+  "license": "Apache-2.0",
+  "main": "dist/index.esm.js",
+  "main:src": "src/index.ts",
+  "types": "src/index.ts",
+  "scripts": {
+    "build": "backstage-cli plugin:build",
+    "lint": "backstage-cli lint",
+    "test": "backstage-cli test",
+    "prepack": "backstage-cli prepack",
+    "postpack": "backstage-cli postpack",
+    "clean": "backstage-cli clean"
+  },
+  "devDependencies": {
+    "@backstage/cli": "^0.1.1-alpha.6",
+    "@types/jest": "^25.2.2",
+    "@types/node": "^12.0.0"
+  },
+  "files": [
+    "dist/**/*.{js,d.ts}"
+  ]
+}
