+++ packages/config/src/types.ts
@@ -14,4 +14,28 @@
  * limitations under the License.
  */
 
-export type AppConfig = any;
+export type JsonObject = { [key in string]: JsonValue };
+export type JsonArray = JsonValue[];
+export type JsonValue =
+  | JsonObject
+  | JsonArray
+  | number
+  | string
+  | boolean
+  | null;
+
+export type AppConfig = JsonObject;
+
+export type Config = {
+  getConfig(key: string): Config;
+
+  getConfigArray(key: string): Config[];
+
+  getNumber(key: string): number | undefined;
+
+  getBoolean(key: string): boolean | undefined;
+
+  getString(key: string): string | undefined;
+
+  getStringArray(key: string): string[] | undefined;
+};
