+++ packages/core-api/src/app/types.ts
@@ -19,6 +19,7 @@ import { IconComponent, SystemIconKey, SystemIcons } from '../icons';
 import { BackstagePlugin } from '../plugin';
 import { ApiHolder } from '../apis';
 import { AppTheme } from '../apis/definitions';
+import { AppConfig } from '@backstage/config';
 
 export type BootErrorPageProps = {
   step: 'load-config';
@@ -30,11 +31,6 @@ export type AppComponents = {
   BootErrorPage: ComponentType<BootErrorPageProps>;
   Progress: ComponentType<{}>;
 };
-
-/**
- * TBD
- */
-export type AppConfig = any;
 
 /**
  * A function that loads in the App config that will be accessible via the ConfigApi.
